+++ tests/cli_current_subcommand_tests.rs
@@ -1,7 +1,9 @@
 mod utils;
 
-use crate::utils::{setup, write_to_file, CURRENT_SCHEME_FILE_NAME};
 use anyhow::Result;
+use utils::{
+    copy_dir_all, setup, write_to_file, CURRENT_SCHEME_FILE_NAME, REPO_DIR, SCHEMES_REPO_NAME,
+};
 
 #[test]
 fn test_cli_current_subcommand_with_setup() -> Result<()> {
@@ -12,7 +14,11 @@ fn test_cli_current_subcommand_with_setup() -> Result<()> {
         setup("test_cli_current_subcommand_with_setup", "current")?;
     let scheme_name = "base16-oceanicnext";
     let current_scheme_path = data_path.join(CURRENT_SCHEME_FILE_NAME);
+    let schemes_dir = data_path.join(format!("{}/{}", REPO_DIR, SCHEMES_REPO_NAME));
+
     write_to_file(&current_scheme_path, scheme_name)?;
+    write_to_file(&current_scheme_path, scheme_name)?;
+    copy_dir_all("./tests/fixtures/schemes", schemes_dir)?;
 
     // ---
     // Act
@@ -22,10 +28,7 @@ fn test_cli_current_subcommand_with_setup() -> Result<()> {
     // ------
     // Assert
     // ------
-    assert!(
-        stdout.contains(scheme_name),
-        "stdout does not contain the expected output"
-    );
+    assert_eq!(stdout, format!("{}\n", scheme_name));
     assert!(
         stderr.is_empty(),
         "stderr does not contain the expected output"
@@ -58,5 +61,75 @@ fn test_cli_current_subcommand_without_setup() -> Result<()> {
         "stderr does not contain the expected output"
     );
 
+    Ok(())
+}
+
+#[test]
+fn test_cli_current_subcommand_with_variant_existing() -> Result<()> {
+    run_test_for_subcommand("variant", |output| output == "dark\n")
+}
+
+#[test]
+fn test_cli_current_subcommand_with_system_existing() -> Result<()> {
+    run_test_for_subcommand("system", |output| output == "base16\n")
+}
+
+#[test]
+fn test_cli_current_subcommand_with_name_existing() -> Result<()> {
+    run_test_for_subcommand("name", |output| output == "Tinty Generated\n")
+}
+
+#[test]
+fn test_cli_current_subcommand_with_slug_existing() -> Result<()> {
+    run_test_for_subcommand("slug", |output| output == "tinty-generated\n")
+}
+
+#[test]
+fn test_cli_current_subcommand_with_description_not_existing() -> Result<()> {
+    run_test_for_subcommand("description", |output| output == "\n")
+}
+
+#[test]
+fn test_cli_current_subcommand_with_author_existing() -> Result<()> {
+    run_test_for_subcommand("author", |output| output == "Tinty\n")
+}
+
+fn run_test_for_subcommand<F>(subcommand: &str, validate_output: F) -> Result<()>
+where
+    F: Fn(&str) -> bool,
+{
+    // -------
+    // Arrange
+    // -------
+    let (_, data_path, command_vec, cleanup) = setup(
+        &format!("test_cli_current_subcommand_with_{}_existing", subcommand),
+        &format!("current {}", subcommand),
+    )?;
+
+    let scheme_name = "base16-tinty-generated";
+    let current_scheme_path = data_path.join(CURRENT_SCHEME_FILE_NAME);
+    let schemes_dir = data_path.join(format!("{}/{}", REPO_DIR, SCHEMES_REPO_NAME));
+
+    write_to_file(&current_scheme_path, scheme_name)?;
+    copy_dir_all("./tests/fixtures/schemes", schemes_dir)?;
+
+    // ---
+    // Act
+    // ---
+    let (stdout, stderr) = utils::run_command(command_vec).unwrap();
+
+    // ------
+    // Assert
+    // ------
+    assert!(
+        validate_output(&stdout), // Pass the actual output to the validation function
+        "stdout does not contain the expected output"
+    );
+    assert!(
+        stderr.is_empty(),
+        "stderr does not contain the expected output"
+    );
+
+    cleanup()?;
     Ok(())
 }